remove TrajectoryInformation to colloid utils
